Yazım hatası düzeltildi

+++ database/migrations/2024_05_19_090541_create_car_models_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('car_models', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('brand_id');
+            $table->unsignedBigInteger('brand_id');
             $table->string('name');
             $table->timestamps();
